Move state files into table
