+++ main.ts
@@ -11,7 +11,7 @@ import { NltSettings, DEFAULT_SETTINGS } from "src/services/settings";
 import { addRow, addColumn } from "src/services/internal/add";
 import { saveAppData } from "src/services/external/save";
 import { createEmptyMarkdownTable } from "src/services/random";
-import { ViewType } from "src/services/state/saveState";
+import { ViewType } from "src/services/settings/saveState";
 
 interface FocusedTable {
 	blockId: string;
