+++ src/components/DateCell/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { CellType } from "src/services/table/types";
-import { isValidCellContent } from "src/services/state/utils";
+import { isValidCellContent } from "src/services/table/utils";
 
 import "./styles.css";
 interface Props {
