+++ src/components/NumberCellEdit/index.tsx
@@ -3,7 +3,7 @@ import { CellType } from "src/services/table/types";
 import {
 	filterNumberFromContent,
 	isValidCellContent,
-} from "src/services/state/utils";
+} from "src/services/table/utils";
 
 import Menu from "../Menu";
