+++ src/services/external/loadUtils.ts
@@ -1,12 +1,12 @@
 import { v4 as uuid } from "uuid";
 import CRC32 from "crc-32";
 
-import { initialHeader, initialRow } from "../state/initialState";
+import { initialHeader, initialRow } from "../table/initialState";
 import { Header, Row, AppData, Cell, Tag } from "../table/types";
 
 import { MARKDOWN_CELLS_REGEX, MARKDOWN_ROWS_REGEX } from "../string/regex";
-import { ViewType } from "../state/saveState";
-import { initialCell } from "../state/initialState";
+import { ViewType } from "../settings/saveState";
+import { initialCell } from "../table/initialState";
 import { getCurrentTimeWithOffset } from "../random";
 
 /**
