--- src/services/state/headerMenu.ts
@@ -1,10 +1,0 @@
-import { CellType } from "src/services/table/types";
-
-export const initialHeaderMenuState = {
-	left: 0,
-	top: 0,
-	id: "",
-	index: 0,
-	content: "",
-	type: CellType.TEXT,
-};
