+++ src/services/table/initialState.ts
@@ -1,7 +1,7 @@
 import { CellType } from "src/services/table/types";
 import { SortDir } from "../sort/types";
 
-import { Cell, Header, Row, Tag } from "../table/types";
+import { Cell, Header, Row, Tag } from "./types";
 
 export const initialHeader = (
 	id: string,
